requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai>=1.51.0
httpx==0.27.2
jinja2==3.1.4
python-dotenv==1.0.1
